+++ src/main/java/htsjdk/samtools/cram/encoding/experimental/GammaIntegerEncoding.java
@@ -15,8 +15,9 @@
  * limitations under the License.
  * ****************************************************************************
  */
-package htsjdk.samtools.cram.encoding;
+package htsjdk.samtools.cram.encoding.experimental;
 
+import htsjdk.samtools.cram.encoding.BitCodec;
 import htsjdk.samtools.cram.io.ExposedByteArrayOutputStream;
 import htsjdk.samtools.cram.io.ITF8;
 import htsjdk.samtools.cram.structure.EncodingID;
@@ -26,12 +27,16 @@ import java.io.InputStream;
 import java.nio.ByteBuffer;
 import java.util.Map;
 
-public class GolombRiceIntegerEncoding implements Encoding<Integer> {
-    private static final EncodingID ENCODING_ID = EncodingID.GOLOMB_RICE;
+public class GammaIntegerEncoding extends ExperimentalEncoding<Integer> {
+    private static final EncodingID ENCODING_ID = EncodingID.GAMMA;
     private int offset;
-    private int m;
 
-    public GolombRiceIntegerEncoding() {
+    public GammaIntegerEncoding() {
+        this(0);
+    }
+
+    public GammaIntegerEncoding(final int offset) {
+        this.offset = offset;
     }
 
     @Override
@@ -39,18 +44,16 @@ public class GolombRiceIntegerEncoding implements Encoding<Integer> {
         return ENCODING_ID;
     }
 
-    public static EncodingParams toParam(final int offset, final int m) {
-        final GolombRiceIntegerEncoding golombRiceIntegerEncoding = new GolombRiceIntegerEncoding();
-        golombRiceIntegerEncoding.offset = offset;
-        golombRiceIntegerEncoding.m = m;
-        return new EncodingParams(ENCODING_ID, golombRiceIntegerEncoding.toByteArray());
+    public static EncodingParams toParam(final int offset) {
+        final GammaIntegerEncoding gammaIntegerEncoding = new GammaIntegerEncoding();
+        gammaIntegerEncoding.offset = offset;
+        return new EncodingParams(ENCODING_ID, gammaIntegerEncoding.toByteArray());
     }
 
     @Override
     public byte[] toByteArray() {
         final ByteBuffer buffer = ByteBuffer.allocate(10);
         ITF8.writeUnsignedITF8(offset, buffer);
-        ITF8.writeUnsignedITF8(m, buffer);
         buffer.flip();
         final byte[] array = new byte[buffer.limit()];
         buffer.get(array);
@@ -59,15 +62,13 @@ public class GolombRiceIntegerEncoding implements Encoding<Integer> {
 
     @Override
     public void fromByteArray(final byte[] data) {
-        final ByteBuffer buffer = ByteBuffer.wrap(data);
-        offset = ITF8.readUnsignedITF8(buffer);
-        m = ITF8.readUnsignedITF8(buffer);
+        offset = ITF8.readUnsignedITF8(data);
     }
 
     @Override
     public BitCodec<Integer> buildCodec(final Map<Integer, InputStream> inputMap,
                                         final Map<Integer, ExposedByteArrayOutputStream> outputMap) {
-        return new GolombRiceIntegerCodec(offset, m);
+        return new GammaIntegerCodec(offset);
     }
 
 }
