+++ src/main/java/htsjdk/samtools/cram/encoding/experimental/GolombLongEncoding.java
@@ -15,8 +15,9 @@
  * limitations under the License.
  * ****************************************************************************
  */
-package htsjdk.samtools.cram.encoding;
+package htsjdk.samtools.cram.encoding.experimental;
 
+import htsjdk.samtools.cram.encoding.BitCodec;
 import htsjdk.samtools.cram.io.ExposedByteArrayOutputStream;
 import htsjdk.samtools.cram.io.ITF8;
 import htsjdk.samtools.cram.structure.EncodingID;
@@ -26,17 +27,12 @@ import java.io.InputStream;
 import java.nio.ByteBuffer;
 import java.util.Map;
 
-public class SubexponentialIntegerEncoding implements Encoding<Integer> {
-    private static final EncodingID ENCODING_ID = EncodingID.SUBEXPONENTIAL;
+public class GolombLongEncoding extends ExperimentalEncoding<Long> {
+    private static final EncodingID ENCODING_ID = EncodingID.GOLOMB;
+    private int m;
     private int offset;
-    private int k;
 
-    public SubexponentialIntegerEncoding() {
-    }
-
-    public SubexponentialIntegerEncoding(final int offset, final int k) {
-        this.offset = offset;
-        this.k = k;
+    public GolombLongEncoding() {
     }
 
     @Override
@@ -44,35 +40,35 @@ public class SubexponentialIntegerEncoding implements Encoding<Integer> {
         return ENCODING_ID;
     }
 
-    public static EncodingParams toParam(final int offset, final int k) {
-        final SubexponentialIntegerEncoding subexponentialIntegerEncoding = new SubexponentialIntegerEncoding();
-        subexponentialIntegerEncoding.offset = offset;
-        subexponentialIntegerEncoding.k = k;
-        return new EncodingParams(ENCODING_ID, subexponentialIntegerEncoding.toByteArray());
+    public static EncodingParams toParam(final int offset, final int m) {
+        final GolombLongEncoding golombLongEncoding = new GolombLongEncoding();
+        golombLongEncoding.offset = offset;
+        golombLongEncoding.m = m;
+        return new EncodingParams(ENCODING_ID, golombLongEncoding.toByteArray());
     }
 
     @Override
     public byte[] toByteArray() {
         final ByteBuffer buffer = ByteBuffer.allocate(10);
         ITF8.writeUnsignedITF8(offset, buffer);
-        ITF8.writeUnsignedITF8(k, buffer);
+        ITF8.writeUnsignedITF8(m, buffer);
         buffer.flip();
-        final byte[] bytes = new byte[buffer.limit()];
-        buffer.get(bytes);
-        return bytes;
+        final byte[] array = new byte[buffer.limit()];
+        buffer.get(array);
+        return array;
     }
 
     @Override
     public void fromByteArray(final byte[] data) {
         final ByteBuffer buffer = ByteBuffer.wrap(data);
         offset = ITF8.readUnsignedITF8(buffer);
-        k = ITF8.readUnsignedITF8(buffer);
+        m = ITF8.readUnsignedITF8(buffer);
     }
 
     @Override
-    public BitCodec<Integer> buildCodec(final Map<Integer, InputStream> inputMap,
-                                        final Map<Integer, ExposedByteArrayOutputStream> outputMap) {
-        return new SubexponentialIntegerCodec(offset, k);
+    public BitCodec<Long> buildCodec(final Map<Integer, InputStream> inputMap,
+                                     final Map<Integer, ExposedByteArrayOutputStream> outputMap) {
+        return new GolombLongCodec(offset, m);
     }
 
 }
